eatgo-customer-api/src/main/java/kr/co/fastcampus/eatgo/interfaces/UserController.java
package kr.co.fastcampus.eatgo.interfaces;

import kr.co.fastcampus.eatgo.application.UserService;
import kr.co.fastcampus.eatgo.domain.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

import java.net.URI;
import java.net.URISyntaxException;

@RestController
public class UserController {
    @Autowired
    private UserService userService;

    @PostMapping("/users")
    public ResponseEntity<?> create(@RequestBody User user) throws URISyntaxException {
        userService.registerUser(user.getEmail(), user.getName(), user.getPassword());

        String uri = "/users/" + user.getId();
        return ResponseEntity.created(new URI(uri)).body("{}");
    }
}
